# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LTT Sales Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
...
  /agents:
    get:
      operationId: listAgents
      tags: [agents]
      summary: List all agents
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: type
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: List of agents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Agent"
    post:
      operationId: createAgent
      tags: [agents]
      summary: Create a new agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAgentBody"
      responses:
        "201":
          description: Created agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
  /agents/bulk-import:
    post:
      operationId: bulkImportAgents
      tags: [agents]
      summary: Bulk import agents from CSV/Excel data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkImportAgentsBody"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkImportResult"
  /agents/{id}:
    get:
      operationId: getAgent
      tags: [agents]
      summary: Get agent by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
        "404":
          description: Not found
    patch:
      operationId: updateAgent
      tags: [agents]
      summary: Update agent
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAgentBody"
      responses:
        "200":
          description: Updated agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
    delete:
      operationId: deleteAgent
      tags: [agents]
      summary: Delete agent
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # INSPECTION REPORTS
  /inspections:
    get:
      operationId: listInspections
      tags: [inspections]
      summary: List inspection reports
      parameters:
        - in: query
          name: agentId
          schema:
            type: integer
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: inspectorId
          schema:
            type: integer
      responses:
        "200":
          description: List of inspection reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InspectionReport"
    post:
      operationId: createInspection
      tags: [inspections]
      summary: Create an inspection report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInspectionBody"
      responses:
        "201":
          description: Created inspection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InspectionReport"
  /inspections/{id}:
    get:
      operationId: getInspection
      tags: [inspections]
      summary: Get inspection report by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Inspection report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InspectionReport"
        "404":
          description: Not found
    patch:
      operationId: updateInspection
      tags: [inspections]
      summary: Update inspection report
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInspectionBody"
      responses:
        "200":
          description: Updated inspection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InspectionReport"

  # INSPECTION SALES LOGS
...
  /inventory:
    get:
      operationId: listInventory
      tags: [inventory]
      summary: List inventory items
      parameters:
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: lowStock
          schema:
            type: boolean
      responses:
        "200":
          description: List of inventory items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
    post:
      operationId: createInventoryItem
      tags: [inventory]
      summary: Create inventory item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInventoryItemBody"
      responses:
        "201":
          description: Created inventory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /inventory/{id}:
    get:
      operationId: getInventoryItem
      tags: [inventory]
      summary: Get inventory item by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Inventory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
        "404":
          description: Not found
    patch:
      operationId: updateInventoryItem
      tags: [inventory]
      summary: Update inventory item
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInventoryItemBody"
      responses:
        "200":
          description: Updated inventory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /inventory/{id}/movement:
    post:
      operationId: recordInventoryMovement
      tags: [inventory]
      summary: Record stock movement (in/out)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryMovementBody"
      responses:
        "200":
          description: Updated inventory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"

  # AGENT SCORES
...
components:
  schemas:
...
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user

    # USERS
    User:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
        role:
          type: string
        department:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - fullName
        - email
        - role
        - isActive
        - createdAt
...
    BulkImportResult:
      type: object
      properties:
        imported:
          type: integer
        skipped:
          type: integer
        errors:
          type: array
          items:
            type: string
      required:
        - imported
        - skipped
        - errors

    # AGENTS
    Agent:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        location:
          type: string
        type:
          type: string
        contractStart:
          type: ["string", "null"]
          format: date-time
        contractEnd:
          type: ["string", "null"]
          format: date-time
        assignedSalesRepId:
          type: ["integer", "null"]
        status:
          type: string
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - location
        - type
        - status
        - createdAt
        - updatedAt
    CreateAgentBody:
      type: object
      properties:
        name:
          type: string
        location:
          type: string
        type:
          type: string
        contractStart:
          type: ["string", "null"]
          format: date-time
        contractEnd:
          type: ["string", "null"]
          format: date-time
        assignedSalesRepId:
          type: ["integer", "null"]
        status:
          type: string
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
      required:
        - name
        - location
        - type
        - status
    UpdateAgentBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        type:
          type: ["string", "null"]
        contractStart:
          type: ["string", "null"]
          format: date-time
        contractEnd:
          type: ["string", "null"]
          format: date-time
        assignedSalesRepId:
          type: ["integer", "null"]
        status:
          type: ["string", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
    BulkImportAgentsBody:
      type: object
      properties:
        agents:
          type: array
          items:
            $ref: "#/components/schemas/CreateAgentBody"
      required:
        - agents

    # INSPECTION REPORTS
    InspectionReport:
      type: object
      properties:
        id:
          type: integer
        agentId:
          type: integer
        inspectorId:
          type: integer
        visitDate:
          type: string
          format: date-time
        location:
          type: ["string", "null"]
        violations:
          type: array
          items:
            type: string
        notes:
          type: ["string", "null"]
        status:
          type: string
        complianceScore:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - agentId
        - inspectorId
        - visitDate
        - status
        - violations
        - createdAt
    CreateInspectionBody:
      type: object
      properties:
        agentId:
          type: integer
        inspectorId:
          type: integer
        visitDate:
          type: string
          format: date-time
        location:
          type: ["string", "null"]
        violations:
          type: array
          items:
            type: string
        notes:
          type: ["string", "null"]
        status:
          type: string
      required:
        - agentId
        - inspectorId
        - visitDate
        - status
    UpdateInspectionBody:
      type: object
      properties:
        violations:
          type: array
          items:
            type: string
        notes:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        complianceScore:
          type: ["number", "null"]

    # SALES LOGS
...
    InventoryItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        quantity:
          type: integer
        minQuantity:
          type: integer
        unit:
          type: string
        serialNumbers:
          type: array
          items:
            type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - category
        - quantity
        - minQuantity
        - unit
        - serialNumbers
        - createdAt
        - updatedAt
    CreateInventoryItemBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        quantity:
          type: integer
        minQuantity:
          type: integer
        unit:
          type: string
        notes:
          type: ["string", "null"]
      required:
        - name
        - category
        - quantity
        - minQuantity
        - unit
    UpdateInventoryItemBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        minQuantity:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
    InventoryMovementBody:
      type: object
      properties:
        type:
          type: string
        quantity:
          type: integer
        reason:
          type: string
      required:
        - type
        - quantity
        - reason

    # AGENT SCORES
...